added a lot of comments in the code to try make sense of what is going on for my and everyones benefit.
Changed the backplane dataobject in backplane_data from resistors to
current_voltage so this makes more sense when the table is generated in
the web browser.

hardware change - adc_cal_module now has both programmable resistors as
20K
backplane.py
Changed the resistance so that both resistor are now 20K rather than the
course being 100K - ln56

Added in calculated & measured ADC_CAL voltages
backplane.py
Changed the self.voltages, self.voltages_raw, self.currents & self.currents_raw arrays to size 15 from 13 to include the two
additional readings. ln-123-126

Added in the code to to measure the "actual values of AUXSAMPLE from the
adc_cal_module.  Hardware mod required to connect ADC to the O/P of the
adc_cal_module. ln-174/5

Added AUXSAMPLE_MEASURED & AUXSAMPLE_SUM to the get_adc_name function.
ln 402

Added the code to calculate the expected adc_cal_module O/P given the
settings of the resistors.  ln 178

Updated the multiplier for the calculated expected value for the new
resistors. ln 253 + 256

updated the max value resistor for fine to 1023. ln - 288

updated the max value for the fine to 80919.3. ln - 284

updated the lines to include true statement and force the
resistor to get the actual value from the hardware. ln 361/2

Things to be done
- update the ad5272 driver to work with the store
function or update default value function so that it will boot up with a
prefferred value.  This is not strictly needed as the values are changed
to the desired values during adc calibration.